requirements.txt
absl-py==2.4.0
annotated-doc==0.0.4
anyio==4.12.1
asttokens==3.0.1
certifi==2026.2.25
cffi==2.0.0
click==8.3.1
cmdkit==2.7.7
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
debugpy==1.8.20
decorator==5.2.1
executing==2.2.1
filelock==3.20.0
fonttools==4.61.1
fsspec==2025.12.0
grpcio==1.78.0
h11==0.16.0
hf-xet==1.3.2
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.5.0
idna==3.11
ipykernel==7.2.0
ipython==9.10.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jupyter_client==8.8.0
jupyter_core==5.9.1
kiwisolver==1.4.9
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.8
matplotlib-inline==0.2.1
mdurl==0.1.2
mpmath==1.3.0
names_generator==0.2.0
nest_asyncio==1.6.0
networkx==3.6.1
numpy==2.3.5
# nvidia-cublas-cu12==12.4.5.8
# nvidia-cuda-cupti-cu12==12.4.127
# nvidia-cuda-nvrtc-cu12==12.4.127
# nvidia-cuda-runtime-cu12==12.4.127
# nvidia-cudnn-cu12==9.1.0.70
# nvidia-cufft-cu12==11.2.1.3
# nvidia-curand-cu12==10.3.5.147
# nvidia-cusolver-cu12==11.6.1.9
# nvidia-cusparse-cu12==12.3.1.170
# nvidia-cusparselt-cu12==0.6.2
# nvidia-nccl-cu12==2.21.5
# nvidia-nvjitlink-cu12==12.4.127
# nvidia-nvtx-cu12==12.4.127
packaging==25.0
parso==0.8.6
pexpect==4.9.0
pillow==12.0.0
pip==26.0.1
platformdirs==4.9.2
prompt_toolkit==3.0.52
protobuf==7.34.0
psutil==7.2.2
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==3.0
Pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
PyYAML==6.0.3
pyzmq==27.1.0
regex==2026.2.28
rich==14.3.3
safetensors==0.7.0
setuptools==80.10.2
shellingham==1.5.4
six==1.17.0
soundfile==0.13.1
stack_data==0.6.3
sympy==1.13.1
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tokenizers==0.22.2
# torch==2.6.0+cu124
# torchaudio==2.6.0+cu124
# torchvision==0.21.0+cu124
tornado==6.5.4
tqdm==4.67.3
traitlets==5.14.3
transformers==5.2.0
# triton==3.2.0
typer==0.24.1
typer-slim==0.24.0
typing_extensions==4.15.0
wcwidth==0.6.0
Werkzeug==3.1.6
wheel==0.46.3
